@@ -1,39 +1,27 @@
-[tool.poetry]
-name = "server"
+[project]
+name = "gadfastapi"
 version = "0.1.0"
-description = ""
-authors = ["Alexander Grishchenko <alexanderdemure@gmail.com>"]
+description = "Add your description here"
 readme = "README.md"
-
-[tool.poetry.dependencies]
-python = "^3.12"
-fastapi = "^0.111.1"
-typer = "^0.12.3"
-uvicorn = "^0.30.3"
-sqlalchemy = "^2.0.31"
-alembic = "^1.13.2"
-asyncpg = "^0.29.0"
-motor = "^3.5.1"
-structlog = "^24.4.0"
-aiologger = "^0.7.0"
-aioboto3 = "^13.1.1"
-pydantic = "^2.8.2"
-pydantic-settings = "^2.3.4"
-jwt = "^1.3.1"
-
-
-[tool.poetry.group.lints.dependencies]
-mypy = "^1.11.0"
-ruff = "^0.5.5"
-isort = "^5.13.2"
-
-
-
-[tool.poetry.group.tests.dependencies]
-pytest = "^8.3.2"
-pytest-asyncio = "^0.23.8"
-radon = "^6.0.1"
-
-[build-system]
-requires = ["poetry-core"]
-build-backend = "poetry.core.masonry.api"
+requires-python = ">=3.13"
+dependencies = [
+    "aioboto3>=14.1.0",
+    "alembic>=1.15.2",
+    "asyncpg>=0.30.0",
+    "fastapi>=0.115.12",
+    "gadlogger>=0.0.1",
+    "gadopenapi>=0.0.1",
+    "gads3client>=0.0.1",
+    "isort>=6.0.1",
+    "motor>=3.7.0",
+    "mypy>=1.15.0",
+    "psycopg2>=2.9.10",
+    "pydantic>=2.11.2",
+    "pydantic-settings>=2.8.1",
+    "python-multipart>=0.0.20",
+    "radon>=6.0.1",
+    "ruff>=0.11.4",
+    "sqlalchemy>=2.0.40",
+    "typer>=0.15.2",
+    "uvicorn>=0.34.0",
+]
