@@ -1,61 +1,69 @@
+from datetime import UTC
 from datetime import datetime
 from typing import Union
 
-from bson import ObjectId
-
+from src.databases import ORM
+from src.databases import ObjectNotFoundError
 from src.databases import Session
-from src.databases import mongo
-from src.databases import postgres
 from src.domain.models import Model
 
 
 class Repository:
-    db: Union[postgres.ORM, mongo.ORM]
+    orm: ORM
     model: Model
+    error: Exception
 
     @classmethod
-    def id(cls, model_id: Union[str, int]) -> Union[str, int, ObjectId]:
-        if issubclass(cls.db, postgres.ORM):
-            return model_id
-        elif issubclass(cls.db, mongo.ORM):
-            return ObjectId(model_id)
-        else:
-            raise NotImplementedError
+    async def one(cls, session: Session, **kwargs) -> "Repository.model":
+        try:
+            row = await cls.orm.one(session, **kwargs)
+        except ObjectNotFoundError:
+            raise cls.error
+
+        return cls.model.from_orm(row)
 
     @classmethod
-    def pk(cls) -> str:
-        if issubclass(cls.db, postgres.ORM):
-            return "id"
-        elif issubclass(cls.db, mongo.ORM):
-            return "_id"
-        else:
-            raise NotImplementedError
+    async def relations(cls, session: Session, **kwargs) -> "Repository.model":
+        try:
+            row = await cls.orm.relations(session, **kwargs)
+        except ObjectNotFoundError:
+            raise cls.error
+
+        return cls.model.from_orm(row)
 
     @classmethod
-    async def get_list_by_ids(cls, session: Session, model_ids: list[Union[str, int]]) -> list["Repository.model"]:
-        rows = await cls.db.get_list(session, field=cls.pk(), value=[cls.id(model_id) for model_id in model_ids])
-        return [cls.model.orm(row) for row in rows]
+    async def paginated(
+        cls,
+        session: Session,
+        filters: dict,
+        sorting: dict,
+        pagination: dict,
+    ) -> tuple[list["Repository.model"], int]:
+        rows, total = await cls.orm.paginated(session, filters, sorting, pagination)
+        return [cls.model.from_orm(row) for row in rows], total
 
     @classmethod
-    async def get_paginated(cls, session: Session, query: dict) -> tuple[list["Repository.model"], int]:
-        rows, total = await cls.db.get_paginated(session, query=query)
-        return [cls.model.orm(row) for row in rows], total
+    async def all(cls, session: Session, **kwargs) -> list["Repository.model"]:
+        rows = await cls.orm.all(session, **kwargs)
+        return [cls.model.from_orm(row) for row in rows]
 
     @classmethod
-    async def create(cls, session: Session, model: dict) -> "Repository.model":
-        row = await cls.db.create(session, model)
-        return cls.model.orm(row)
+    async def count(cls, session: Session, **kwargs) -> int:
+        return await cls.orm.count(session, **kwargs)
 
     @classmethod
-    async def update(cls, session: Session, model_id: Union[str, int], **kwargs) -> None:
-        await cls.db.update(
-            session=session,
-            field=cls.pk(),
-            value=cls.id(model_id),
-            updated_at=datetime.utcnow(),
-            **kwargs,
-        )
+    async def exists(cls, session: Session, **kwargs) -> bool:
+        return await cls.orm.exists(session, **kwargs)
 
     @classmethod
-    async def delete(cls, session: Session, model_id: Union[str, int]) -> None:
-        return await cls.db.delete(session, field=cls.pk(), value=cls.id(model_id))
+    async def create(cls, session: Session, model: Model) -> "Repository.model":
+        row = await cls.orm.create(session, row=model.dict())
+        return cls.model.from_orm(row)
+
+    @classmethod
+    async def update(cls, session: Session, id: Union[str, int], **kwargs) -> None:
+        await cls.orm.update(session=session, id=id, updated_at=datetime.now(UTC), **kwargs)
+
+    @classmethod
+    async def delete(cls, session: Session, **kwargs) -> None:
+        return await cls.orm.delete(session, **kwargs)
