@@ -13,7 +13,7 @@ from src.framework import settings
 # this is the Alembic Config object, which provides
 # access to the values within the .ini file in use.
 config = context.config
-config.set_main_option("sqlalchemy.url", str(settings.POSTGRES_URI))
+config.set_main_option("sqlalchemy.url", settings.POSTGRES_URL)
 
 # Interpret the config file for Python logging.
 # This line sets up loggers basically.
