@@ -1,7 +1,6 @@
 from typing import TypeVar
 from typing import Union
 
-from src.databases import mongo
 from src.databases import postgres
 
-Session = TypeVar("Session", bound=Union[mongo.Session, postgres.Session])
+Session = TypeVar("Session", bound=Union[postgres.Session])
