@@ -1,19 +1,14 @@
+from typing import Self
+
+from src.utils import strings
+
 from .base import Model
 
 
 class Dummy(Model):
-    id: int
+    id: int | None = None
+    name: str
 
     @classmethod
-    def init(cls) -> dict:
-        return dict()
-
-    def jwt(self) -> None: ...
-
-
-class DummyDocument(Model):
-    id: str
-
-    @classmethod
-    def init(cls) -> dict:
-        return dict()
+    def init(cls, name: str) -> Self:
+        return cls(name=strings.to_lower(name))
